feat: AI Gateway OSS/EE split — removable ee/ tree + OSS leak scrub (#33960) (#33961)
Split llm/gateway into an OSS core + an ee/ subtree that is physically removed when
producing the OSS snapshot (folder-removal via .oss-exclude, mirroring the llm-server
ee/ pattern — no build tags). The whole-module exclude is gone; only ee/ + the bridge
are excluded.

Mechanism:
- Registration hooks with OSS defaults (mirror common.RegisterDatabaseManagerHook):
  routing.RegisterRuleLoader (default: static config only), ratelimit.RegisterBuilder
  (default: nil = allow-all; Limiter.Enabled() is nil-safe), proxy.RegisterCredResolver
  (default: operator/account credential), metering.RegisterBodyLogSink (default: no-op).
- EE impls self-register via init() from ee/{routing,ratelimit,secrets,bodylog};
  cmd/gateway/imports_enterprise.go blank-imports them (excluded from OSS).
- main.go builds every dep from the hooks — imports no ee/ package.

Leak scrub (OSS source must not reveal the paid tier or its capabilities):
- Neutralized all OSS comments (no enterprise/EE/OSS/PHI/DLP/BYO/distributed/budget
  references); extension points read as generic plugin seams with no-op defaults.
- Excluded llm/gateway/docs/ (design docs describe the full architecture).
- Moved secret decryption (common/secrets.go) into ee/secrets/crypto.go.
- Renamed the no-op `phi` stage to a neutral `filter`.

Verified: build+test with ee/ present; OSS build compiles with ee/ + bridge deleted;
check-oss-boundary.sh passes; leak-grep over OSS source is empty; gofmt + vet clean.

Fixes #33960
