+++ llm/gateway/ee/secrets/crypto.go
@@ -1,4 +1,4 @@
-package common
+package secrets
 
 import (
 	"crypto/aes"
@@ -9,10 +9,9 @@ import (
 	"nudgebee/llm-gateway/config"
 )
 
-// Decrypt reverses llm-server's Encrypt (AES-GCM, 12-byte nonce prefix, hex-encoded)
-// using the shared NB encryption key. The gateway only needs to READ per-tenant
-// provider secrets from integration_config_values, so only Decrypt is ported.
-func Decrypt(encrypted string) (string, error) {
+// decrypt reverses llm-server's Encrypt (AES-GCM, 12-byte nonce prefix, hex-encoded)
+// using the shared NB encryption key, so a stored provider secret can be read back.
+func decrypt(encrypted string) (string, error) {
 	data, err := hex.DecodeString(encrypted)
 	if err != nil {
 		return "", err
